検証監査 修正バッチ1: core/ecs/gameframework の UAF・オーバーフロー・無限ループ
全 src/ の検証監査 (stub / 致命バグ / nullcheck 漏れ) で確定した高〜中リスクを修正。

[core]
- Log.cpp: 初期化を二重ロック化 (g_init_lock 取得 → 準備完了後に g_inited を
  publish)、Shutdown は g_inited を先に落としてから VirtualFree (race/UAF 回避)。
  FormatTimestamp の i64 オーバーフロー (約25時間で破綻) を分割計算で回避。
  drop カウンタを別管理し DroppedCount のゼロ化バグを解消。
- Json.cpp: _strtod_l に渡す locale が NULL の場合 strtod へフォールバック。
- threading/ThreadPool.h: Done() を CAS ループ化し 0 未満アンダーフローを防止。
- threading/JobGraph.cpp: カウンタを submit 成功ごとに +1、失敗時に巻き戻す方式へ
  (一括 Add の早期完了・二重 Done を排除)。

[ecs]
- Query.h: Each/EachParallel が primary の dense[] をローカルへスナップショットして
  から反復。fn が構造的変更で SparseSet を再確保しても dangling しない。

[gameframework]
- 系統的 UAF (コールバック/再確保で TArray の生ポインタ・参照が dangling) を是正:
  PickupSystem / CinematicsDirector / Sequence / WeaponSystem / CraftingSystem /
  SceneCommandQueue / SceneEventBus。値コピー後にコールバック、再取得 or
  size 再ガードを徹底。
- Random.cpp: RangeInt の (max-min) を u32 昇格後に減算し符号付きオーバーフロー UB を回避。
- Settings.cpp: AppendF32 で v>=2^64 の f32→u64 範囲外変換 UB を回避 (double で桁展開)。
- DialogueScript.cpp: RunUntilBlocked に Jump ラベル循環ガードを追加し無限ループを防止。

検証: hello_persist_verify === ALL PASS ===、フルビルド green。

+++ acs/src/gameframework/Random.cpp
@@ -99,8 +99,11 @@ i32 FRandom::RangeInt(i32 min, i32 max) noexcept {
         const i32 t = min; min = max; max = t;
     }
     // span = max - min + 1, 1 以上が保証される。
-    // (max - min) は最大 2^31-1-(-2^31)=2^32-1 で u32 に収まるよう u32 で計算。
-    const u32 span = static_cast<u32>(max - min) + 1u;
+    // (max - min) を i32 のまま計算すると min=INT32_MIN, max=INT32_MAX で
+    // 2^32-1 となり i32 の表現域を超えて符号付きオーバーフロー (UB)。
+    // 各オペランドを先に u32 へ昇格してから減算する (u32 の減算は modulo 2^32 で
+    // well-defined、結果も正しい差になる)。
+    const u32 span = (static_cast<u32>(max) - static_cast<u32>(min)) + 1u;
     if (span == 0u) {
         // span オーバーフロー (min=INT32_MIN, max=INT32_MAX)。
         // フル 32bit 範囲を返してよい。
